CRUD for card, category and transaction

+++ controllers/card.js
@@ -3,67 +3,69 @@ const cardService = require('../services/api/card');
 
 
 const handleGetCardById = async (req, res) => {
-    const { params } = req;
+    const { params, account: { id: accountId }, user: { id: userId } } = req;
     try {
 
-        const card = await cardService.getCardById({ params });
+        const card = await cardService.getCardById({ params, accountId, userId });
         res.status(201).json(card);
 
     } catch (error) {
-        console.log(`Error in handleGetCardById: ${error}`);
+        console.log(`Error in handleGetCardById: ${error} for user: ${userId} and account: ${accountId}`);
         res.status(400).json({ error: error.message });
     }
 }
 
 
 const handleGetAllCards = async (req, res) => {
+    const { account: { id: accountId }, user: { id: userId } } = req;
     try {
-
-        const cards = await cardService.getAllCards();
+        
+        const cards = await cardService.getAllCards({ accountId, userId });
         res.status(201).json(cards);
 
     } catch (error) {
-        console.log(`Error in handleGetAllCards: ${error}`);
+        console.log(`Error in handleGetAllCards: ${error} for user: ${userId} and account: ${accountId}`);
         res.status(400).json({ error: error.message });
     }
 }
 
 const hanldeCardCreation = async (req, res) => {
-    const { body } = req;
+    const { body, account: { id: accountId }, user: { id: userId } } = req;
     try {
 
-        const card = await cardService.createCard({ body });
+        const card = await cardService.createCard({ body, accountId, userId });
         res.status(201).json(card);
 
     } catch (error) {
-        console.log(`Error in hanldeCardCreation: ${error}`);
+        console.log(`Error in hanldeCardCreation: ${error} for user: ${userId} and account: ${accountId}`);
         res.status(400).json({ error: error.message });
     }
 }
 
 
 const handleCardUpdation = async (req, res) => {
-    const { body } = req;
+    const { body, account: { id: accountId }, params: { cardId} } = req;
     try {
 
-        const card = await cardService.updateCard({ body });
+        const card = await cardService.updateCard({ body, accountId, cardId });
         res.status(201).json(card);
 
     } catch (error) {
-        console.log(`Error in handleCardUpdation: ${error}`);
+        console.log(`Error in handleCardUpdation: ${error} for account: ${accountId}`);
         res.status(400).json({ error: error.message });
     }
 }
 
 
 const handleCardDeletion = async (req, res) => {
+    const { account: { id: accountId }, params } = req;
     try {
 
-        const card = await cardService.deleteCard();
+        const card = await cardService.deleteCard({ accountId, params });
         res.status(201).json(card);
 
     } catch (error) {
-        console.log(`Error in handleCardDeletion: ${error}`);
+        console.log(`Error in handleCardDeletion: ${error} for account: ${accountId}`);
         res.status(400).json({ error: error.message });
     }
 }
